+++ CHANGELOG.md
@@ -22,6 +22,10 @@ Also possible, but for this project less relevant, is `Deprecated` for soon-to-b
 ### Added
 * 5-to-2 reactions for NNbar annihilations via the stochastic collision criterion
 * 2-to-5 reactions for NNbar annihilations to 5 pions
+* thermodynamic lattice output
+
+### Changed
+* computation of thermodynamic quantities optionally restriced to participants only
 
 
 ## [SMASH-2.0.2](https://github.com/smash-transport/smash/compare/SMASH-2.0.1...2.0.2)